LayoutTests/media/media-webm-opus-error.html

<!DOCTYPE html>
<html>
    <head>
        <script src="video-test.js"></script>
        <script>
            var audio = null;
            async function start() {
                audio = mediaElement = document.getElementsByTagName('audio')[0];

                testExpected("audio.canPlayType('audio/webm; codecs=opus')", "probably");
                waitForEvent('error', endTest);
                run('audio.src = "content/opus_variable_witherror.webm"');
                await waitFor(audio, 'loadedmetadata');
                testExpected('audio.duration', '0.98', '>=');
            }
        </script>
    </head>
    <body onload="start()">
        <audio controls></audio>
        <p>Check that a webm with garbage content at the end generate an error.</p>
    </body>
</html>
